Bump frontend dependencies to newer versions
We stay on React 15 because it is the last version to have addons.
Some of the functionality of React 14 has been outlined into separate
modules; some of the internal modules have been moved.
Since raven-js 2.0.0, the functionality of the "native" plugin is
provided out-of-the-box, and the plugin itself is removed.