Revise emoji display
The trademark sign test is obsolete as newer twemoji versions do not
capture the character in the first place. Translating anonymous emoji
has been explicitly allowed to be able to display emoji that are in
twemoji but not in emoji-annotation-to-unicode; the long-term fix is,
of course, to fix the latter to include the new emoji.