Remove obsolete remark
The code has been very stable (in fact, it received no changes
whatsoever) for several years already, so most of the remark is
obsolete. As an exception, the note about being slow to merge
complex pull request is correct (but understates being slow to
merge any pull requests at all).