Fix staff-invade + mock KMS corrupting manager key
staff-invade is the only place where the StaffUnlock() method of room
manager key objects is used. In the PSQL backend's implementation of
that method, a shallow copy of the KeyEncryptingKey was formerly passed
to the KMS' DecryptKey() method. The mock KMS' DecryptKey() assumes that
it owns the passed key's ciphertext's backing array and modifies that
array in-place. However, when called by the removed code, the KEK's
ciphertext slice aliased (due to the shallow copy) that of a long-lived
copy of the KEK, which became corrupted. Any later uses of the impacted
room's ManagerKey would operate on copies of the corrupted KEK and fail.

The new code performs a deep copy of the KEK (which seems to be the
intended mode of usage) and thus avoids any aliasing issues.

This bug did probably not arise with the AWS KMS because it likely does
not modify the supplied slices' backing arrays in-place.