Sort client/lib/*.js into subdirectories
This is part of a multi-patch effort to make client/lib/ less of a
dumping ground.

The only JS files not moved are "entry points", or scripts closely
related to entry point scripts. The other scripts are moved into util/
if they are reasonably general-purpose (including MessageData.js, which
despite its name is *very* generic) and into heim/ otherwise.

Previously, client/lib/heim/ held only the low-level socket module. If
there was any separation of "Heim-specific logic" intended by that, it
is not apparent to me in the rest of the code.

The import statements in dependent files are deliberately not sorted
in order to make the diff more manageable.