Rework room switcher URL selection
The previous mechanism (viz. submitting a GET form) had the aesthetic
blemish of attaching a question mark to the new tab's URL. This has
caused a user enough discomfort to launch a PR to change the behavior.

The new approach explicitly computes the new room's URL and, as a
drive-by enhancement, carries over the current page's query and
fragment. The latter is particularly relevant to the backend-less
frontend development server (which uses "hash flags" to connect to
a fully-featured backend).

Closes #3.