Add same-tab option to room switcher
Now, the room switcher can also be used to open another room in the same
tab.

Opening a new tab from the keyboard is broken: React (in its deep folly)
invokes the keyboard event handler asynchronously, which means that the
tab opening happens outside of a user interaction context. This will be
fixed in a follow-up commit.